+++ changelog.md
@@ -19,6 +19,7 @@
 - Fixed some errors with the field MultiSelect in DataView
 - Fixed missing usage of section callbacks in DataView
 - Fixed the missing usage of the label for field Button in DataView
+- Fixed security of values in "One" method
 
 ## [3.0.2] - 02.08.2026